@@ -1,19 +1,44 @@
 {
   "private": true,
-  "name": "tgui",
-  "version": "4.1.0",
+  "name": "tgui-workspace",
+  "version": "4.3.0",
   "workspaces": [
     "packages/*"
   ],
-  "scripts": {
-    "build": "eslint packages && cd packages/tgui && npx webpack --mode=production",
-    "watch": "cd packages/tgui-dev-server && node --experimental-modules index.js",
-    "analyze": "cd packages/tgui && npx webpack --mode=production --env.analyze=1",
-    "lint": "eslint packages"
-  },
   "dependencies": {
-    "babel-eslint": "^10.0.3",
-    "eslint": "^7.4.0",
-    "eslint-plugin-react": "^7.17.0"
+    "@babel/core": "^7.13.15",
+    "@babel/eslint-parser": "^7.13.14",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-transform-jscript": "^7.12.13",
+    "@babel/preset-env": "^7.13.15",
+    "@babel/preset-typescript": "^7.13.0",
+    "@types/jest": "^26.0.22",
+    "@types/jsdom": "^16.2.10",
+    "@types/node": "^14.14.41",
+    "@typescript-eslint/parser": "^4.22.0",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-inferno": "^6.2.0",
+    "babel-plugin-transform-remove-console": "^6.9.4",
+    "common": "workspace:*",
+    "css-loader": "^5.2.2",
+    "eslint": "^7.24.0",
+    "eslint-plugin-radar": "^0.2.1",
+    "eslint-plugin-react": "^7.23.2",
+    "file-loader": "^6.2.0",
+    "inferno": "^7.4.8",
+    "jest": "^26.6.3",
+    "jest-circus": "^26.6.3",
+    "jsdom": "^16.5.3",
+    "mini-css-extract-plugin": "^1.4.1",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
+    "style-loader": "^2.0.0",
+    "terser-webpack-plugin": "^5.1.1",
+    "typescript": "^4.2.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.33.2",
+    "webpack-bundle-analyzer": "^4.4.1",
+    "webpack-cli": "^4.6.0"
   }
 }
