@@ -26,7 +26,9 @@ export const createStore = (reducer, enhancer) => {
 
   const dispatch = action => {
     currentState = reducer(currentState, action);
-    listeners.forEach(fn => fn());
+    for (let i = 0; i < listeners.length; i++) {
+      listeners[i]();
+    }
   };
 
   // This creates the initial store by causing each reducer to be called
@@ -81,7 +83,7 @@ export const applyMiddleware = (...middlewares) => {
 export const combineReducers = reducersObj => {
   const keys = Object.keys(reducersObj);
   let hasChanged = false;
-  return (prevState, action) => {
+  return (prevState = {}, action) => {
     const nextState = { ...prevState };
     for (let key of keys) {
       const reducer = reducersObj[key];
@@ -96,4 +98,56 @@ export const combineReducers = reducersObj => {
       ? nextState
       : prevState;
   };
+};
+
+/**
+ * A utility function to create an action creator for the given action
+ * type string. The action creator accepts a single argument, which will
+ * be included in the action object as a field called payload. The action
+ * creator function will also have its toString() overriden so that it
+ * returns the action type, allowing it to be used in reducer logic that
+ * is looking for that action type.
+ *
+ * @param {string} type The action type to use for created actions.
+ * @param {any} prepare (optional) a method that takes any number of arguments
+ * and returns { payload } or { payload, meta }. If this is given, the
+ * resulting action creator will pass it's arguments to this method to
+ * calculate payload & meta.
+ *
+ * @public
+ */
+export const createAction = (type, prepare = null) => {
+  const actionCreator = (...args) => {
+    if (!prepare) {
+      return { type, payload: args[0] };
+    }
+    const prepared = prepare(...args);
+    if (!prepared) {
+      throw new Error('prepare function did not return an object');
+    }
+    const action = { type };
+    if ('payload' in prepared) {
+      action.payload = prepared.payload;
+    }
+    if ('meta' in prepared) {
+      action.meta = prepared.meta;
+    }
+    return action;
+  };
+  actionCreator.toString = () => '' + type;
+  actionCreator.type = type;
+  actionCreator.match = action => action.type === type;
+  return actionCreator;
+};
+
+
+// Implementation specific
+// --------------------------------------------------------
+
+export const useDispatch = context => {
+  return context.store.dispatch;
+};
+
+export const useSelector = (context, selector) => {
+  return selector(context.store.getState());
 };
