@@ -1,4 +1,10 @@
 /**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
+/**
  * Removes excess whitespace and indentation from the string.
  */
 export const multiline = str => {
@@ -91,7 +97,7 @@ export const toTitleCase = str => {
     return str;
   }
   // Handle string
-  const WORDS_UPPER = ['Id', 'Tv', 'Rcd'];
+  const WORDS_UPPER = ['Id', 'Tv', 'Rcd']; // VOREStation Edit
   const WORDS_LOWER = [
     'A', 'An', 'And', 'As', 'At', 'But', 'By', 'For', 'For', 'From', 'In',
     'Into', 'Near', 'Nor', 'Of', 'On', 'Onto', 'Or', 'The', 'To', 'With',
