@@ -1,7 +1,15 @@
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
 import { clamp } from 'common/math';
 import { pureComponentHooks } from 'common/react';
 import { Component, createRef } from 'inferno';
 import { AnimatedNumber } from './AnimatedNumber';
+
+const DEFAULT_UPDATE_RATE = 400;
 
 /**
  * Reduces screen offset to a single number based on the matrix provided.
@@ -66,7 +74,7 @@ export class DraggableControl extends Component {
         if (dragging && onDrag) {
           onDrag(e, value);
         }
-      }, 500);
+      }, this.props.updateRate || DEFAULT_UPDATE_RATE);
       document.addEventListener('mousemove', this.handleDragMove);
       document.addEventListener('mouseup', this.handleDragEnd);
     };
@@ -165,12 +173,12 @@ export class DraggableControl extends Component {
       unit,
       minValue,
       maxValue,
+      unclamped,
       format,
       onChange,
       onDrag,
       children,
       // Input props
-      forcedInputWidth,
       height,
       lineHeight,
       fontSize,
@@ -207,7 +215,6 @@ export class DraggableControl extends Component {
         style={{
           display: !editing ? 'none' : undefined,
           height: height,
-          width: forcedInputWidth,
           'line-height': lineHeight,
           'font-size': fontSize,
         }}
@@ -215,7 +222,22 @@ export class DraggableControl extends Component {
           if (!editing) {
             return;
           }
-          const value = clamp(e.target.value, minValue, maxValue);
+          let value;
+          if (unclamped) {
+            value = parseFloat(e.target.value);
+          }
+          else {
+            value = clamp(
+              parseFloat(e.target.value),
+              minValue,
+              maxValue);
+          }
+          if (Number.isNaN(value)) {
+            this.setState({
+              editing: false,
+            });
+            return;
+          }
           this.setState({
             editing: false,
             value,
@@ -230,7 +252,22 @@ export class DraggableControl extends Component {
         }}
         onKeyDown={e => {
           if (e.keyCode === 13) {
-            const value = clamp(e.target.value, minValue, maxValue);
+            let value;
+            if (unclamped) {
+              value = parseFloat(e.target.value);
+            }
+            else {
+              value = clamp(
+                parseFloat(e.target.value),
+                minValue,
+                maxValue);
+            }
+            if (Number.isNaN(value)) {
+              this.setState({
+                editing: false,
+              });
+              return;
+            }
             this.setState({
               editing: false,
               value,
