@@ -1,3 +1,9 @@
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
 import { classes } from 'common/react';
 import { Component } from 'inferno';
 import { Box } from './Box';
@@ -41,17 +47,18 @@ export class Dropdown extends Component {
   }
 
   buildMenu() {
-    const { options = [], placeholder } = this.props;
+    const { options = [], placeholder } = this.props; // VOREStation edit
     const ops = options.map(option => (
-      <div
+      <Box
         key={option}
         className="Dropdown__menuentry"
         onClick={() => {
           this.setSelected(option);
         }}>
         {option}
-      </div>
+      </Box>
     ));
+    // VOREStation addition start
     if (placeholder) {
       ops.unshift((
         <div
@@ -64,22 +71,26 @@ export class Dropdown extends Component {
         </div>
       ));
     }
-    return ops;
+    // VOREStation addition end
+    return ops.length ? ops : 'No Options Found';
   }
 
   render() {
     const { props } = this;
     const {
+      icon,
+      iconRotation,
+      iconSpin,
       color = 'default',
       over,
       noscroll,
       nochevron,
       width,
-      maxHeight,
       onClick,
       selected,
       disabled,
-      placeholder,
+      displayText,
+      placeholder, // VOREStation Addition
       ...boxProps
     } = props;
     const {
@@ -95,7 +106,6 @@ export class Dropdown extends Component {
         tabIndex="-1"
         style={{
           'width': width,
-          'max-height': maxHeight,
         }}
         className={classes([
           noscroll && 'Dropdown__menu-noscroll' || 'Dropdown__menu',
@@ -123,8 +133,15 @@ export class Dropdown extends Component {
             }
             this.setOpen(!this.state.open);
           }}>
+          {icon && (
+            <Icon
+              name={icon}
+              rotation={iconRotation}
+              spin={iconSpin}
+              mr={1} />
+          )}
           <span className="Dropdown__selected-text">
-            {this.state.selected || placeholder}
+            {displayText ? displayText : (this.state.selected || placeholder) /* VOREStation Edit */ }
           </span>
           {!!nochevron || (
             <span className="Dropdown__arrow-button">
