@@ -21,10 +21,10 @@ export const Knob = props => {
   const {
     // Draggable props (passthrough)
     animated,
-    forcedInputWidth,
     format,
     maxValue,
     minValue,
+    unclamped,
     onChange,
     onDrag,
     step,
@@ -38,10 +38,9 @@ export const Knob = props => {
     fillValue,
     color,
     ranges = {},
-    size,
+    size = 1,
     bipolar,
     children,
-    popUpPosition,
     ...rest
   } = props;
   return (
@@ -49,10 +48,10 @@ export const Knob = props => {
       dragMatrix={[0, -1]}
       {...{
         animated,
-        forcedInputWidth,
         format,
         maxValue,
         minValue,
+        unclamped,
         onChange,
         onDrag,
         step,
@@ -82,7 +81,7 @@ export const Knob = props => {
         const effectiveColor = color
           || keyOfMatchingRange(fillValue ?? value, ranges)
           || 'default';
-        const rotation = (scaledDisplayValue - 0.5) * 270;
+        const rotation = Math.min((scaledDisplayValue - 0.5) * 270, 225);
         return (
           <div
             className={classes([
@@ -94,7 +93,7 @@ export const Knob = props => {
             ])}
             {...computeBoxProps({
               style: {
-                'font-size': size + 'rem',
+                'font-size': size + 'em',
                 ...style,
               },
               ...rest,
@@ -110,10 +109,7 @@ export const Knob = props => {
               </div>
             </div>
             {dragging && (
-              <div className={classes([
-                'Knob__popupValue',
-                popUpPosition && 'Knob__popupValue--' + popUpPosition,
-              ])}>
+              <div className="Knob__popupValue">
                 {displayElement}
               </div>
             )}
@@ -133,8 +129,8 @@ export const Knob = props => {
                 className="Knob__ringFill"
                 style={{
                   'stroke-dashoffset': (
-                    ((bipolar ? 2.75 : 2.00) - scaledFillValue * 1.5)
-                      * Math.PI * 50
+                    Math.max(((bipolar ? 2.75 : 2.00) - scaledFillValue * 1.5)
+                      * Math.PI * 50, 0)
                   ),
                 }}
                 cx="50"
