@@ -1,3 +1,9 @@
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
 import { classes } from 'common/react';
 import { computeBoxClassName, computeBoxProps } from './Box';
 import { Dimmer } from './Dimmer';
@@ -6,9 +12,10 @@ export const Modal = props => {
   const {
     className,
     children,
-    onEnter,
+    onEnter, // VOREStation Addition
     ...rest
   } = props;
+  // VOREStation Addition Start
   let handleKeyDown;
   if (onEnter) {
     handleKeyDown = e => {
@@ -18,9 +25,10 @@ export const Modal = props => {
       }
     };
   }
+  // VOREStation Addition End
   return (
     <Dimmer
-      onKeyDown={handleKeyDown}>
+      onKeyDown={handleKeyDown /* VOREStation Edit */}>
       <div
         className={classes([
           'Modal',
