@@ -10,6 +10,8 @@ import { Component, createRef } from 'inferno';
 import { AnimatedNumber } from './AnimatedNumber';
 import { Box } from './Box';
 
+const DEFAULT_UPDATE_RATE = 400;
+
 export class NumberInput extends Component {
   constructor(props) {
     super(props);
@@ -64,7 +66,7 @@ export class NumberInput extends Component {
         if (dragging && onDrag) {
           onDrag(e, value);
         }
-      }, 500);
+      }, this.props.updateRate || DEFAULT_UPDATE_RATE);
       document.addEventListener('mousemove', this.handleDragMove);
       document.addEventListener('mouseup', this.handleDragEnd);
     };
@@ -215,7 +217,16 @@ export class NumberInput extends Component {
             if (!editing) {
               return;
             }
-            const value = clamp(e.target.value, minValue, maxValue);
+            const value = clamp(
+              parseFloat(e.target.value),
+              minValue,
+              maxValue);
+            if (Number.isNaN(value)) {
+              this.setState({
+                editing: false,
+              });
+              return;
+            }
             this.setState({
               editing: false,
               value,
@@ -230,7 +241,16 @@ export class NumberInput extends Component {
           }}
           onKeyDown={e => {
             if (e.keyCode === 13) {
-              const value = clamp(e.target.value, minValue, maxValue);
+              const value = clamp(
+                parseFloat(e.target.value),
+                minValue,
+                maxValue);
+              if (Number.isNaN(value)) {
+                this.setState({
+                  editing: false,
+                });
+                return;
+              }
               this.setState({
                 editing: false,
                 value,
