@@ -1,11 +1,19 @@
-import { classes } from 'common/react';
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
+import { canRender, classes } from 'common/react';
 import { computeBoxClassName, computeBoxProps } from './Box';
-import { Button } from './Button';
+import { Icon } from './Icon';
 
 export const Tabs = props => {
   const {
     className,
     vertical,
+    fill,
+    fluid,
     children,
     ...rest
   } = props;
@@ -16,13 +24,13 @@ export const Tabs = props => {
         vertical
           ? 'Tabs--vertical'
           : 'Tabs--horizontal',
+        fill && 'Tabs--fill',
+        fluid && 'Tabs--fluid',
         className,
         computeBoxClassName(rest),
       ])}
       {...computeBoxProps(rest)}>
-      <div className="Tabs__tabBox">
-        {children}
-      </div>
+      {children}
     </div>
   );
 };
@@ -31,20 +39,42 @@ const Tab = props => {
   const {
     className,
     selected,
-    altSelection,
+    color,
+    icon,
+    leftSlot,
+    rightSlot,
+    children,
     ...rest
   } = props;
   return (
-    <Button
+    <div
       className={classes([
-        'Tabs__tab',
-        selected && 'Tabs__tab--selected',
-        altSelection && selected && 'Tabs__tab--altSelection',
+        'Tab',
+        'Tabs__Tab',
+        'Tab--color--' + color,
+        selected && 'Tab--selected',
         className,
+        ...computeBoxClassName(rest),
       ])}
-      selected={!altSelection && selected}
-      color="transparent"
-      {...rest} />
+      {...computeBoxProps(rest)}>
+      {canRender(leftSlot) && (
+        <div className="Tab__left">
+          {leftSlot}
+        </div>
+      ) || !!icon && (
+        <div className="Tab__left">
+          <Icon name={icon} />
+        </div>
+      )}
+      <div className="Tab__text">
+        {children}
+      </div>
+      {canRender(rightSlot) && (
+        <div className="Tab__right">
+          {rightSlot}
+        </div>
+      )}
+    </div>
   );
 };
 
