@@ -1,4 +1,11 @@
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
 export { AnimatedNumber } from './AnimatedNumber';
+export { Blink } from './Blink';
 export { BlockQuote } from './BlockQuote';
 export { Box } from './Box';
 export { Button } from './Button';
@@ -13,17 +20,23 @@ export { Dropdown } from './Dropdown';
 export { Flex } from './Flex';
 export { Grid } from './Grid';
 export { Icon } from './Icon';
+export { InfinitePlane } from './InfinitePlane';
 export { Input } from './Input';
 export { Knob } from './Knob';
 export { LabeledControls } from './LabeledControls';
 export { LabeledList } from './LabeledList';
-export { NanoMap } from './NanoMap';
 export { Modal } from './Modal';
+export { NanoMap } from './NanoMap';
 export { NoticeBox } from './NoticeBox';
 export { NumberInput } from './NumberInput';
 export { ProgressBar } from './ProgressBar';
+export { Popper } from './Popper';
+export { RoundGauge } from './RoundGauge';
 export { Section } from './Section';
 export { Slider } from './Slider';
+export { Stack } from './Stack';
 export { Table } from './Table';
 export { Tabs } from './Tabs';
+export { TextArea } from './TextArea';
+export { TimeDisplay } from './TimeDisplay';
 export { Tooltip } from './Tooltip';
