@@ -25,6 +25,11 @@ const SI_SYMBOLS = [
   'E', // exa
   'Z', // zetta
   'Y', // yotta
+  'R', // ronna
+  'Q', // quecca
+  'F',
+  'N',
+  'H',
 ];
 
 const SI_BASE_INDEX = SI_SYMBOLS.indexOf(' ');
@@ -95,6 +100,104 @@ export const formatMoney = (value, precision = 0) => {
   return result;
 };
 
+/**
+ * Formats a floating point number as a number on the decibel scale.
+ */
+export const formatDb = value => {
+  const db = 20 * Math.log(value) / Math.log(10);
+  const sign = db >= 0 ? '+' : '–';
+  let formatted = Math.abs(db);
+  if (formatted === Infinity) {
+    formatted = 'Inf';
+  }
+  else {
+    formatted = toFixed(formatted, 2);
+  }
+  return sign + formatted + ' dB';
+};
+
+const SI_BASE_TEN_UNIT = [
+  '',
+  '· 10³', // kilo
+  '· 10⁶', // mega
+  '· 10⁹', // giga
+  '· 10¹²', // tera
+  '· 10¹⁵', // peta
+  '· 10¹⁸', // exa
+  '· 10²¹', // zetta
+  '· 10²⁴', // yotta
+  '· 10²⁷', // ronna
+  '· 10³⁰', // quecca
+  '· 10³³',
+  '· 10³⁶',
+  '· 10³⁹',
+];
+
+const SI_BASE_TEN_INDEX = SI_BASE_TEN_UNIT.indexOf(' ');
+
+
+/**
+ * Formats a number to a human readable form, by reducing it to SI units.
+ * TODO: This is quite a shit code and shit math, needs optimization.
+ */
+export const formatSiBaseTenUnit = (
+  value,
+  minBase1000 = -SI_BASE_TEN_INDEX,
+  unit = ''
+) => {
+  if (typeof value !== 'number' || !Number.isFinite(value)) {
+    return value;
+  }
+  const realBase10 = Math.floor(Math.log10(value));
+  const base10 = Math.floor(Math.max(minBase1000 * 3, realBase10));
+  const realBase1000 = Math.floor(realBase10 / 3);
+  const base1000 = Math.floor(base10 / 3);
+  const symbolIndex = clamp(
+    SI_BASE_TEN_INDEX + base1000,
+    0,
+    SI_BASE_TEN_UNIT.length);
+  const symbol = SI_BASE_TEN_UNIT[symbolIndex];
+  const scaledNumber = value / Math.pow(1000, base1000);
+  const scaledPrecision = realBase1000 > minBase1000
+    ? (2 + base1000 * 3 - base10)
+    : 0;
+  // TODO: Make numbers bigger than precision value show
+  // up to 2 decimal numbers.
+  const finalString = (
+    toFixed(scaledNumber, scaledPrecision)
+    + ' ' + symbol + ' ' + unit
+  );
+  return finalString.trim();
+};
+
+/**
+ * Formats decisecond count into HH::MM::SS display by default
+ * "short" format does not pad and adds hms suffixes
+ */
+export const formatTime = (val, formatType) => {
+  // THERE IS AS YET INSUFFICIENT DATA FOR A MEANINGFUL ANSWER
+  // HH:MM:SS
+  // 00:02:13
+  const seconds = toFixed(Math.floor((val/10) % 60));
+  const minutes = toFixed(Math.floor((val/(10*60)) % 60));
+  const hours = toFixed(Math.floor((val/(10*60*60)) % 24));
+  switch (formatType) {
+    case "short": {
+      const hours_truncated = hours > 0 ? `${hours}h` : "";
+      const minutes_truncated = minutes > 0 ? `${minutes}m` : "";
+      const seconds_truncated = seconds > 0 ? `${seconds}s` : "";
+      return `${hours_truncated}${minutes_truncated}${seconds_truncated}`;
+    }
+    default: {
+      const seconds_padded = seconds.padStart(2, "0");
+      const minutes_padded = minutes.padStart(2, "0");
+      const hours_padded = hours.padStart(2, "0");
+      return `${hours_padded}:${minutes_padded}:${seconds_padded}`;
+    }
+  }
+};
+
+/* VOREStation Addition Start */
 export const formatCommaNumber = value => {
   if (!Number.isFinite(value)) {
     return value;
@@ -104,13 +207,4 @@ export const formatCommaNumber = value => {
   parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",");
   return parts.join(".");
 };
-
-// Function from https://stackoverflow.com/a/34841026. CC BY-SA 4.0.
-export const formatTime = seconds => {
-  let hours = Math.floor(seconds / 3600);
-  let minutes = Math.floor(seconds / 60) % 60;
-  return [hours, minutes, seconds % 60]
-    .map(v => v < 10 ? "0" + v : v)
-    .filter((v, i) => v !== "00" || i > 0)
-    .join(":");
-};
+/* VOREStation Addition End */
