@@ -81,11 +81,9 @@ const ChemDispenserChemicals = (properties, context) => {
         direction="row"
         wrap="wrap"
         height="100%"
-        spacingPrecise="2"
-        align="flex-start"
-        alignContent="flex-start">
+        align="flex-start">
         {chemicals.map((c, i) => (
-          <Flex.Item key={i} grow="1" basis="40%" height="20px">
+          <Flex.Item key={i} grow="1" m={0.2} basis="40%" height="20px">
             <Button
               icon="arrow-circle-down"
               width="100%"
