@@ -29,19 +29,19 @@ const viewRecordModalBodyOverride = (modal, context) => {
         <LabeledList.Item label="Damage">
           {damages.length > 1 ? (
             <Fragment>
-              <Box color={COLORS.damageType.oxy} display="inline">
+              <Box color={COLORS.damageType.oxy} inline>
                 {damages[0]}
               </Box>
               &nbsp;|&nbsp;
-              <Box color={COLORS.damageType.toxin} display="inline">
+              <Box color={COLORS.damageType.toxin} inline>
                 {damages[2]}
               </Box>
               &nbsp;|&nbsp;
-              <Box color={COLORS.damageType.brute} display="inline">
+              <Box color={COLORS.damageType.brute} inline>
                 {damages[3]}
               </Box>
               &nbsp;|&nbsp;
-              <Box color={COLORS.damageType.burn} display="inline">
+              <Box color={COLORS.damageType.burn} inline>
                 {damages[1]}
               </Box>
             </Fragment>
@@ -197,7 +197,7 @@ const CloningConsoleMain = (props, context) => {
         level="2"
         buttons={
           <Fragment>
-            <Box display="inline" color="label">
+            <Box inline color="label">
               Scanner Lock:&nbsp;
             </Box>
             <Button
@@ -308,7 +308,7 @@ const CloningConsoleMain = (props, context) => {
                 <Box
                   bold
                   color={pod.biomass >= 150 ? "good" : "bad"}
-                  display="inline">
+                  inline>
                   <Icon
                     name={pod.biomass >= 150 ? "circle" : "circle-o"}
                   />&nbsp;
@@ -409,7 +409,7 @@ const CloningConsoleStatus = (props, context) => {
         <Fragment>
           {!!autoallowed && (
             <Fragment>
-              <Box display="inline" color="label">
+              <Box inline color="label">
                 Auto-processing:&nbsp;
               </Box>
               <Button
