@@ -350,7 +350,7 @@ const PhoneTab = (props, context) => {
   return (
     <Section title="Phone">
       <LabeledList>
-        <LabeledList.Item label="Target EPv2 Address" verticalAlign="middle"> 
+        <LabeledList.Item label="Target EPv2 Address" verticalAlign="middle">
           <Flex align="center">
             <Flex.Item grow={1}>
               <Input
@@ -464,7 +464,7 @@ const PhoneTab = (props, context) => {
           {!!invitesSent.length && (
             <LabeledList>
               {invitesSent.map(invite => (
-                <LabeledList.Item label={decodeHtmlEntities(invite.name)} key={invite.address}> 
+                <LabeledList.Item label={decodeHtmlEntities(invite.name)} key={invite.address}>
                   <Box>{decodeHtmlEntities(invite.address)}</Box>
                   <Box>
                     <Button
@@ -764,7 +764,7 @@ const MessagingThreadTab = (props, context) => {
             icon="eye"
             selected={clipboardMode}
             tooltip="Exit Clipboard Mode"
-            tooltipPosition="bottom-left"
+            tooltipPosition="bottom-end"
             onClick={() => setClipboardMode(!clipboardMode)} />
         }
         height="100%"
@@ -807,7 +807,7 @@ const MessagingThreadTab = (props, context) => {
           icon="eye"
           selected={clipboardMode}
           tooltip="Enter Clipboard Mode"
-          tooltipPosition="bottom-left"
+          tooltipPosition="bottom-end"
           onClick={() => setClipboardMode(!clipboardMode)} />
       }
       height="100%"
@@ -843,7 +843,7 @@ TabToTemplate[MESSSUBTAB] = <MessagingThreadTab />;
 /* News */
 const NewsTab = (props, context) => {
   const { act, data } = useBackend(context);
-  
+
   const {
     feeds,
     target_feed,
@@ -1021,7 +1021,7 @@ const WeatherTab = (props, context) => {
         {!!weather.length && (
           <LabeledList>
             {weather.map(wr => (
-              <LabeledList.Item label={wr.Planet} key={wr.Planet}> 
+              <LabeledList.Item label={wr.Planet} key={wr.Planet}>
                 <LabeledList>
                   <LabeledList.Item label="Time">
                     {wr.Time}
