@@ -230,8 +230,8 @@ const MedicalRecordsViewGeneral = (_properties, context) => {
         <LabeledList>
           {general.fields.map((field, i) => (
             <LabeledList.Item key={i} label={field.field}>
-              <Box height="20px" display="inline-block">
-                {field.value.split("\n").map(m => <Box key={m}>{m}</Box>)}
+              <Box height="20px" display="inline-block" preserveWhitespace>
+                {field.value}
               </Box>
               {!!field.edit && (
                 <Button
@@ -293,8 +293,8 @@ const MedicalRecordsViewMedical = (_properties, context) => {
         {medical.fields.map((field, i) => (
           <LabeledList.Item
             key={i}
-            label={field.field}>
-            {field.value.split("\n").map(m => <Box key={m}>{m}</Box>)}
+            label={field.field} preserveWhitespace>
+            {field.value}
             <Button
               icon="pen"
               ml="0.5rem"
@@ -312,7 +312,7 @@ const MedicalRecordsViewMedical = (_properties, context) => {
         )
           : medical.comments.map((comment, i) => (
             <Box key={i}>
-              <Box color="label" display="inline">
+              <Box color="label" inline>
                 {comment.header}
               </Box><br />
               {comment.text}
