@@ -86,7 +86,7 @@ const SleeperOccupant = (props, context) => {
       title="Occupant"
       buttons={(
         <Fragment>
-          <Box color="label" display="inline">
+          <Box color="label" inline>
             Auto-eject if dead:&nbsp;
           </Box>
           <Button
