@@ -126,9 +126,10 @@ const VoreBellySelectionAndCustomization = (props, context) => {
         {our_bellies.map(belly => (
           <Tabs.Tab
             key={belly.name}
-            color={belly.selected ? "green" : digestModeToColor[belly.digest_mode]}
+            selected={belly.selected}
+            textColor={digestModeToColor[belly.digest_mode]}
             onClick={() => act("bellypick", { bellypick: belly.ref })}>
-            <Box inline color={belly.selected && digestModeToColor[belly.digest_mode] || null}>
+            <Box inline textColor={belly.selected && digestModeToColor[belly.digest_mode] || null}>
               {belly.name} ({belly.contents})
             </Box>
           </Tabs.Tab>
@@ -165,7 +166,9 @@ const VoreSelectedBelly = (props, context) => {
     nutrition_percent,
     digest_brute,
     digest_burn,
+    digest_oxy,
     bulge_size,
+    display_absorbed_examine,
     shrink_grow_size,
     emote_time,
     emote_active,
@@ -270,6 +273,9 @@ const VoreSelectedBelly = (props, context) => {
               onClick={() => act("set_attribute", { attribute: "b_msgs", msgtype: "em" })}
               content="Examine Message (when full)" />
             <Button
+              onClick={() => act("set_attribute", { attribute: "b_msgs", msgtype: "ema" })}
+              content="Examine Message (with absorbed victims)" />
+            <Button
               onClick={() => act("set_attribute", { attribute: "b_msgs", msgtype: "im_digest" })}
               content="Idle Messages (Digest)" />
             <Button
@@ -303,6 +309,11 @@ const VoreSelectedBelly = (props, context) => {
                 <Button
                   onClick={() => act("set_attribute", { attribute: "b_burn_dmg" })}
                   content={digest_burn} />
+              </LabeledList.Item>
+              <LabeledList.Item label="Digest Suffocation Damage">
+                <Button
+                  onClick={() => act("set_attribute", { attribute: "b_oxy_dmg" })}
+                  content={digest_oxy} />
               </LabeledList.Item>
               <LabeledList.Item label="Nutritional Gain">
                 <Button
@@ -390,6 +401,13 @@ const VoreSelectedBelly = (props, context) => {
                 <Button
                   onClick={() => act("set_attribute", { attribute: "b_bulge_size" })}
                   content={bulge_size * 100 + "%"} />
+              </LabeledList.Item>
+              <LabeledList.Item label="Display Absorbed Examines">
+                <Button
+                  onClick={() => act("set_attribute", { attribute: "b_display_absorbed_examine" })}
+                  icon={display_absorbed_examine ? "toggle-on" : "toggle-off"}
+                  selected={display_absorbed_examine}
+                  content={display_absorbed_examine ? "True" : "False"} />
               </LabeledList.Item>
               <LabeledList.Item label="Shrink/Grow Size">
                 <Button
@@ -601,6 +619,7 @@ const VoreUserPreferences = (props, context) => {
     show_vore_fx,
     can_be_drop_prey,
     can_be_drop_pred,
+    allow_spontaneous_tf,
     step_mechanics_active,
     pickup_mechanics_active,
     noisy,
@@ -761,7 +780,7 @@ const VoreUserPreferences = (props, context) => {
                 + " Click this to enable showing FX.")}
             content={show_vore_fx ? "Show Vore FX" : "Do Not Show Vore FX"} />
         </Flex.Item>
-        <Flex.Item basis="49%">
+        <Flex.Item basis="32%">
           <Button
             onClick={() => act("toggle_leaveremains")}
             icon={digest_leave_remains ? "toggle-on" : "toggle-off"}
@@ -773,9 +792,9 @@ const VoreUserPreferences = (props, context) => {
               + "if they do not, they will not leave your remains behind, even with this on. Click to disable remains"
               : ("Regardless of Predator Setting, you will not leave remains behind."
                 + " Click this to allow leaving remains.")}
-            content={digest_leave_remains ? "Allow Leaving Remains Behind" : "Do Not Allow Leaving Remains Behind"} />
+            content={digest_leave_remains ? "Allow Leaving Remains" : "Do Not Allow Leaving Remains"} />
         </Flex.Item>
-        <Flex.Item basis="49%">
+        <Flex.Item basis="32%" grow={1}>
           <Button
             onClick={() => act("toggle_pickuppref")}
             icon={pickup_mechanics_active ? "toggle-on" : "toggle-off"}
@@ -788,6 +807,19 @@ const VoreUserPreferences = (props, context) => {
               : ("You will not participate in pick-up mechanics."
                 + " Click this to allow picking up/being picked up.")}
             content={pickup_mechanics_active ? "Pick-up Mechanics Enabled" : "Pick-up Mechanics Disabled"} />
+        </Flex.Item>
+        <Flex.Item basis="32%">
+          <Button
+            onClick={() => act("toggle_allow_spontaneous_tf")}
+            icon={allow_spontaneous_tf ? "toggle-on" : "toggle-off"}
+            selected={allow_spontaneous_tf}
+            fluid
+            tooltip={"This toggle is for spontaneous or environment related transformation"
+            + " as a victim, such as via chemicals. "
+            + (allow_spontaneous_tf
+              ? "Click here to allow being spontaneously transformed."
+              : "Click here to disable being spontaneously transformed.")}
+            content={allow_spontaneous_tf ? "Spontaneous TF Enabled" : "Spontaneous TF Disabled"} />
         </Flex.Item>
         <Flex.Item basis="49%">
           <Button
