@@ -102,7 +102,7 @@ export const ComplexModal = (props, context) => {
         placeholder="ENTER to submit"
         width="100%"
         my="0.5rem"
-        autofocus
+        autoFocus
         onChange={(_e, val) => {
           curValue = val;
         }}
@@ -188,7 +188,7 @@ export const ComplexModal = (props, context) => {
       maxHeight={props.maxHeight || (window.innerHeight / 2 + "px")}
       onEnter={modalOnEnter}
       mx="auto">
-      <Box display="inline">
+      <Box inline>
         {text}
       </Box>
       {modalBody}
