@@ -9,7 +9,7 @@ export const FullscreenNotice = (props, context) => {
     title = 'Welcome',
   } = props;
   return (
-    <Section title={title} height="100%" stretchContents>
+    <Section title={title} height="100%" fill>
       <Flex height="100%" align="center" justify="center">
         <Flex.Item textAlign="center" mt="-2rem">
           {children}
