@@ -7,7 +7,6 @@
 import { resolveAsset } from '../assets';
 import { useBackend } from '../backend';
 import { Box, Button } from '../components';
-import { refocusLayout } from './Layout';
 import { Window } from './Window';
 
 export const NtosWindow = (props, context) => {
@@ -15,7 +14,6 @@ export const NtosWindow = (props, context) => {
     title,
     width = 575,
     height = 700,
-    resizable,
     theme = 'ntos',
     children,
   } = props;
@@ -36,14 +34,9 @@ export const NtosWindow = (props, context) => {
       title={title}
       width={width}
       height={height}
-      theme={theme}
-      resizable={resizable}>
+      theme={theme}>
       <div className="NtosWindow">
-        <div
-          className="NtosWindow__header NtosHeader"
-          onMouseDown={() => {
-            refocusLayout();
-          }}>
+        <div className="NtosWindow__header NtosHeader">
           <div className="NtosHeader__left">
             <Box inline bold mr={2}>
               {PC_stationtime}
@@ -68,13 +61,11 @@ export const NtosWindow = (props, context) => {
                   src={resolveAsset(PC_ntneticon)} />
               )}
             </Box>
-            {!!PC_showbatteryicon && PC_batteryicon && (
+            {!!(PC_showbatteryicon && PC_batteryicon) && (
               <Box inline mr={1}>
-                {PC_batteryicon && (
-                  <img
-                    className="NtosHeader__icon"
-                    src={resolveAsset(PC_batteryicon)} />
-                )}
+                <img
+                  className="NtosHeader__icon"
+                  src={resolveAsset(PC_batteryicon)} />
                 {PC_batterypercent && (
                   PC_batterypercent
                 )}
@@ -90,6 +81,7 @@ export const NtosWindow = (props, context) => {
             {!!PC_showexitprogram && (
               <Button
                 width="26px"
+                lineHeight="22px"
                 textAlign="center"
                 color="transparent"
                 icon="window-minimize-o"
@@ -101,11 +93,12 @@ export const NtosWindow = (props, context) => {
               <Button
                 mr="-3px"
                 width="26px"
+                lineHeight="22px"
                 textAlign="center"
                 color="transparent"
                 icon="window-close-o"
                 tooltip="Close"
-                tooltipPosition="bottom-left"
+                tooltipPosition="bottom-start"
                 onClick={() => act('PC_exit')} />
             )}
             {!PC_showexitprogram && (
@@ -117,7 +110,7 @@ export const NtosWindow = (props, context) => {
                 color="transparent"
                 icon="power-off"
                 tooltip="Power off"
-                tooltipPosition="bottom-left"
+                tooltipPosition="bottom-start"
                 onClick={() => act('PC_shutdown')} />
             )}
           </div>
