@@ -4,6 +4,7 @@
  * @license MIT
  */
 
-export { Layout, refocusLayout } from './Layout';
+export { refocusLayout, Layout } from './Layout';
 export { NtosWindow } from './NtosWindow';
+export { Pane } from './Pane';
 export { Window } from './Window';
